services/provider/requirements.txt
fastapi==0.122.0    
uvicorn==0.38.0
psycopg2-binary==2.9.11
pydantic==2.12.5